+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.1.2
 Release TBD
 - Add customizable template header for notifications.
+- Add admin access to toggle user's preference to receive cards.
 
 ## Version 1.1.1
 Release 2021-12-16